feat: textinput component instead of separate input and textarea
