+++ packages/frontend/src/modules/app/components/TodoList/TodoList.styles.ts
@@ -1,5 +1,5 @@
 import { css } from '@emotion/css';
-import { THEME, COLORS } from '../../../../shared/styles/constants';
+import { THEME, COLORS } from '~shared/styles/constants';
 
 export const emptyStateStyles = css`
 	text-align: center;
