+++ packages/frontend/src/modules/app/pages/NewTodoPage/NewTodoPage.styles.ts
@@ -1,5 +1,5 @@
 import { css } from '@emotion/css';
-import { THEME } from '../../../../shared/styles/constants';
+import { THEME } from '~shared/styles/constants';
 
 export const checkboxContainerStyles = css`
 	display: flex;
