--- packages/frontend/src/shared/components/textarea/textarea.styles.ts
@@ -1,8 +1,0 @@
-import { css } from '@emotion/css';
-import { inputStyles } from '../input/input.styles';
-
-export const textareaStyles = css`
-	${inputStyles}
-	min-height: 100px;
-	resize: vertical;
-`;
